+++ packages/photos/client/views/panels.js
@@ -1,6 +1,0 @@
-
-Template.panelSettings_photos.helpers({
-	setAvatar: function() {
-		return K.Profile.setAvatar;
-	}
-});