+++ packages/upload/package.js
@@ -21,6 +21,7 @@ Package.onUse(function(api) {
     'i18n/de.js',
     'i18n/es.js',
     'i18n/fr.js',	
+    'i18n/sv.js', 
   ]);
 
   api.addFiles([