+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 - generalize plugin upload inputFile_upload template
 - new admin method updateUserPassword
 - new admin settings settings.admin.emailOnNewUser
+- plugin photos has two new method resizePhoto, updateAvatar
+- plugin photos new settinngs mimeFileType to check image format
 
 ## v1.5.3
 - new settings to disable user markers settings.public.map.layerUsers.enabled